Added a function to calculate the level, row and col
level is calculated by divinding the y by 200 to get it in a range from 0-3. The colum is calucalted by getting the y value at the current level in a range of 0-1 and changing the width of the board + the offset at that given y level. The row is currently not being calculated as the row's are weirdly offset