Added input to get Level, Row and Col (#2)
* Added std::bitset board, added functionality for making a move

* Added checks for wins in all directions, added TEMPORARY cin input for testing, removed unnecessary bufferMask

* Added a function to calculate the level, row and col

level is calculated by divinding the y by 200 to get it in a range from 0-3. The colum is calucalted by getting the y value at the current level in a range of 0-1 and changing the width of the board + the offset at that given y level. The row is currently not being calculated as the row's are weirdly offset

* Added row checking into the input function

* Added input handler to tidy up code

* Drawing tokens started | need to decide on sprite

* Added texture manager, refactored input class to work based on dynamic board size

Co-authored-by: ben-millar <c00236772@itcarlow.ie>